Move MultibuttonWidget to lib/widget, add ButtonAlignment
